Modifed CalculatorPageViewModel to accept additional injected dependencies; added mock of TipDatabase so that we can setup additional unit testing of the CalculatorPageViewmodel, implemented the FileHelper class for the Android project.

+++ tipclac_core/Interfaces/ITipCalcTransaction.cs
@@ -4,7 +4,7 @@ namespace tipcalc_core.Interfaces
 {
     public interface ITipCalcTransaction
     {
-        decimal Id { get; set; }
+        int Id { get; set; }
 
         decimal Total { get; set; }
         decimal Tip { get; set; }
